+++ src/oscar/Graphics/Materials.h
@@ -1,3 +1,4 @@
 #pragma once
 
 #include <oscar/Graphics/Materials/MeshBasicMaterial.h>
+#include <oscar/Graphics/Materials/MeshPhongMaterial.h>